libusbhost: improve usb_device_open handling
This method is used whenever Android tries to access a device using the
USB host API.

I found a situation where this method was not able to actually provide
access to a USB device, even if the USB device was present.

Using a portable DAC (Sabaj DA3), this happens:
- DAC is plugged in to the phone,
- on Linux, the USB device nodes are created in '/dev/bus/usb/00X/00Y'
- with that specific DAC, 3 nodes are created: 001, 002 and 003
- node 002, for some reason, disappears almost immediately after its
  creation. The actual node that we need is 003
- due to how the usb_device_open method was managing the situation,
  there was a one (1) second timeout before actually failing and
returning
- before returning, the libusbhost was still actually calling the 'open'
  method on a file that didn't exist anymore.

The fix for this is pretty simple: just check if the file exists as
first thing. If it doesn't, return immediately without doing anything
else.

While we are there, also move the 'sleep' before all the checks: if we
are waiting for the permissions to be correctly set, it makes sense to
wait before checking them. Otherwise, the first verification will most
likely fail.

Change-Id: Ic2d3a4c0f6d2d806b9b214f519d14f7e83e709b4
Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>